refactor: explicitly import error type from errors in file-systems
