+++ src/file-systems/__mocks__/__common__.ts
@@ -1,13 +1,13 @@
 const fileSystemMock = {
   listFiles: jest.fn()
     .mockReturnValue([])
-    .mockName("FileSystem.listFiles"),
+    .mockName("file-systems.FileSystem.listFiles"),
   readFile: jest.fn()
     .mockReturnValue(["", null])
-    .mockName("FileSystem.readFile"),
+    .mockName("file-systems.FileSystem.readFile"),
   writeFile: jest.fn()
     .mockReturnValue(null)
-    .mockName("FileSystem.writeFile"),
+    .mockName("file-systems.FileSystem.writeFile"),
 };
 
 export {
